requirements.txt
google-generativeai
python-dotenv
pdfplumber
Pillow
fastapi
fastapi[standard]
uvicorn
pydantic